(PC-17806) feat(auth): move QuitSignupModal to pages folders
