+++ src/features/auth/pages/signup/QuitSignupModal/QuitSignupModal.native.test.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 
-import { QuitSignupModal } from 'features/auth/components/QuitSignupModal'
 import { SignupStep } from 'features/auth/enums'
 import { navigateToHome } from 'features/navigation/helpers'
 import { analytics } from 'libs/firebase/analytics'
 import { fireEvent, render } from 'tests/utils'
+
+import { QuitSignupModal } from './QuitSignupModal'
 
 jest.mock('features/navigation/helpers')
