+++ src/features/auth/pages/signup/SignupForm.tsx
@@ -4,9 +4,9 @@ import React, { FunctionComponent, useState, useCallback } from 'react'
 import { Keyboard } from 'react-native'
 
 import { useSignUp } from 'features/auth/api/useSignUp'
-import { QuitSignupModal } from 'features/auth/components/QuitSignupModal'
 import { SIGNUP_NUMBER_OF_STEPS } from 'features/auth/constants'
 import { PreValidationSignupStep } from 'features/auth/enums'
+import { QuitSignupModal } from 'features/auth/pages/signup/QuitSignupModal/QuitSignupModal'
 import { PreValidationSignupStepProps, SignupData } from 'features/auth/types'
 import { RootStackParamList } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
